requirements.txt
asgiref==3.5.2
Django==3.2.14
pytz==2022.2.1
recursos-humanos==0.1
sqlparse==0.4.3
psycopg2==2.9.4



django-extensions==3.2.1

Pillow==9.3.0
mod_wsgi==4.9.4

djangorestframework==3.14.0
WhiteNoise==6.2.0





